+++ packages/openapi/test/out/samples/pets/post.js
@@ -1,0 +1,7 @@
+fetch("http://localhost:8080/pets", {
+  body: JSON.stringify({
+    "id": 0,
+    "name": "string",
+    "tag": "string"
+  })
+});