+++ packages/create-app/CHANGELOG.md
@@ -1,5 +1,9 @@
 # create-next-docs-app
 
+## 13.3.2
+
+## 13.3.1
+
 ## 13.3.0
 
 ### Patch Changes